# src/swagger.json
{
	"components": {
...
		"schemas": {
			"ProfileType": {
				"type": "string",
				"enum": [
					"stakeholder",
					"admin",
					"unit owner"
				]
			},
			"RecordStatus": {
				"type": "string",
				"enum": [
					"active",
					"inactive"
				]
			},
...
			"CommunityAttr": {
				"properties": {
					"description": {
						"type": "string"
					}
				},
				"required": [
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChargeType": {
				"type": "string",
				"enum": [
					"unit",
					"percentage",
					"amount"
				]
			},
			"PostingType": {
				"type": "string",
				"enum": [
					"monthly",
					"manual",
					"accrued",
					"interest",
					"quarterly"
				]
			},
			"ChargeAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"communityId": {
						"type": "number",
						"format": "double"
					},
					"community": {
						"$ref": "#/components/schemas/CommunityAttr"
					},
					"code": {
						"type": "string"
					},
					"rate": {
						"type": "number",
						"format": "double"
					},
					"chargeType": {
						"$ref": "#/components/schemas/ChargeType"
					},
					"postingType": {
						"$ref": "#/components/schemas/PostingType"
					},
					"thresholdInMonths": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"priority": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"passOn": {
						"type": "boolean",
						"nullable": true
					}
				},
				"required": [
					"communityId",
					"code",
					"rate",
					"chargeType",
					"postingType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TransactionType": {
				"type": "string",
				"enum": [
					"charged",
					"collected"
				]
			},
...
			"PaymentType": {
				"type": "string",
				"enum": [
					"cash",
					"check"
				]
			},
			"ProfileAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"mobileNumber": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/ProfileType"
					},
					"status": {
						"$ref": "#/components/schemas/RecordStatus"
					},
					"scopes": {
						"type": "string"
					},
					"remarks": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"username",
					"password",
					"email",
					"type",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DisbursementAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"voucherId": {
						"type": "number",
						"format": "double"
					},
					"purchaseOrderId": {
						"type": "number",
						"format": "double"
					},
					"chargeId": {
						"type": "number",
						"format": "double"
					},
					"releasedBy": {
						"type": "number",
						"format": "double"
					},
					"paymentType": {
						"$ref": "#/components/schemas/PaymentType"
					},
					"details": {
						"type": "string"
					},
					"checkNumber": {
						"type": "string"
					},
					"checkPostingDate": {
						"type": "string",
						"format": "date-time"
					},
					"checkIssuingBank": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"releasedByProfile": {
						"$ref": "#/components/schemas/ProfileAttr"
					},
					"charge": {
						"$ref": "#/components/schemas/ChargeAttr"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"chargeId",
					"releasedBy",
					"paymentType",
					"details",
					"amount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FieldError": {
				"properties": {
					"value": {
						"type": "string",
						"nullable": true
					},
					"field": {
						"type": "string",
						"nullable": true
					},
					"type": {
						"type": "string",
						"nullable": true
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"value",
					"field",
					"type",
					"message"
				],
				"type": "object"
			},
			"EntityError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"fieldErrors": {
						"items": {
							"$ref": "#/components/schemas/FieldError"
						},
						"type": "array",
						"default": []
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PropertyAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "string"
					},
					"floorArea": {
						"type": "number",
						"format": "double"
					},
					"address": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/RecordStatus"
					}
				},
				"required": [
					"code",
					"floorArea",
					"address",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaymentDetailAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"collectedBy": {
						"type": "number",
						"format": "double"
					},
					"orNumber": {
						"type": "string"
					},
					"paymentType": {
						"$ref": "#/components/schemas/PaymentType"
					},
					"checkNumber": {
						"type": "string"
					},
					"checkPostingDate": {
						"type": "string",
						"format": "date-time"
					},
					"checkIssuingBank": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"collectedBy",
					"orNumber",
					"paymentType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TransactionAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"chargeId": {
						"type": "number",
						"format": "double"
					},
					"charge": {
						"$ref": "#/components/schemas/ChargeAttr"
					},
					"propertyId": {
						"type": "number",
						"format": "double"
					},
					"property": {
						"$ref": "#/components/schemas/PropertyAttr"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"transactionPeriod": {
						"type": "string",
						"format": "date-time"
					},
					"transactionType": {
						"$ref": "#/components/schemas/TransactionType"
					},
					"waivedBy": {
						"type": "number",
						"format": "double"
					},
					"comments": {
						"type": "string"
					},
					"paymentDetailId": {
						"type": "number",
						"format": "double"
					},
					"paymentDetail": {
						"$ref": "#/components/schemas/PaymentDetailAttr"
					},
					"rateSnapshot": {
						"type": "number",
						"format": "double"
					},
					"batchId": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"chargeId",
					"propertyId",
					"amount",
					"transactionPeriod",
					"transactionType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PropertyAssignmentAttr": {
				"properties": {
					"profileId": {
						"type": "number",
						"format": "double"
					},
					"propertyId": {
						"type": "number",
						"format": "double"
					},
					"profile": {
						"$ref": "#/components/schemas/ProfileAttr"
					},
					"property": {
						"$ref": "#/components/schemas/PropertyAttr"
					}
				},
				"required": [
					"profileId",
					"propertyId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaymentHistoryView": {
				"properties": {
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"collectedBy": {
						"type": "string"
					},
					"checkIssuingBank": {
						"type": "string"
					},
					"checkPostingDate": {
						"type": "string",
						"format": "date-time"
					},
					"checkNumber": {
						"type": "string"
					},
					"paymentType": {
						"$ref": "#/components/schemas/PaymentType"
					},
					"orNumber": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"transactionPeriod": {
						"type": "string",
						"format": "date-time"
					},
					"amount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"createdAt",
					"collectedBy",
					"paymentType",
					"orNumber",
					"code",
					"transactionPeriod",
					"amount"
				],
				"type": "object"
			},
			"PropertyTransactionHistoryView": {
				"properties": {
					"paymentHistory": {
						"items": {
							"$ref": "#/components/schemas/PaymentHistoryView"
						},
						"type": "array"
					},
					"transactionHistory": {
						"items": {
							"$ref": "#/components/schemas/TransactionAttr"
						},
						"type": "array"
					},
					"previousBalance": {
						"type": "number",
						"format": "double"
					},
					"targetYear": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"paymentHistory",
					"transactionHistory",
					"previousBalance",
					"targetYear"
				],
				"type": "object"
			},
			"RequestStatus": {
				"type": "string",
				"enum": [
					"approved",
					"rejected",
					"pending",
					"cancelled"
				]
			},
			"ExpenseAttr": {
				"properties": {
					"voucherId": {
						"type": "number",
						"format": "double"
					},
					"purchaseRequestId": {
						"type": "number",
						"format": "double"
					},
					"purchaseOrderId": {
						"type": "number",
						"format": "double"
					},
					"categoryId": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"unitCost": {
						"type": "number",
						"format": "double"
					},
					"totalCost": {
						"type": "number",
						"format": "double"
					},
					"waivedBy": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"categoryId",
					"category",
					"description",
					"quantity",
					"unitCost",
					"totalCost"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PurchaseOrderAttr": {
				"properties": {
					"series": {
						"type": "string"
					},
					"isClosed": {
						"type": "boolean"
					},
					"approverProfiles": {
						"items": {
							"$ref": "#/components/schemas/ProfileAttr"
						},
						"type": "array"
					},
					"rejectedByProfile": {
						"$ref": "#/components/schemas/ProfileAttr"
					},
					"requestedByProfile": {
						"$ref": "#/components/schemas/ProfileAttr"
					},
					"disbursements": {
						"items": {
							"$ref": "#/components/schemas/DisbursementAttr"
						},
						"type": "array"
					},
					"expenses": {
						"items": {
							"$ref": "#/components/schemas/ExpenseAttr"
						},
						"type": "array"
					},
					"comments": {
						"type": "string",
						"nullable": true
					},
					"rejectedBy": {
						"type": "number",
						"format": "double"
					},
					"approvedBy": {
						"type": "string",
						"nullable": true
					},
					"requestedDate": {
						"type": "string",
						"format": "date-time"
					},
					"requestedBy": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/RequestStatus"
					},
					"totalCost": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"chargeId": {
						"type": "number",
						"format": "double"
					},
					"otherDetails": {
						"type": "string",
						"nullable": true
					},
					"fulfillmentDate": {
						"type": "string",
						"format": "date-time"
					},
					"vendorName": {
						"type": "string"
					},
					"purchaseRequestId": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"series",
					"requestedDate",
					"requestedBy",
					"status",
					"totalCost",
					"description",
					"chargeId",
					"fulfillmentDate",
					"vendorName",
					"purchaseRequestId"
				],
				"type": "object"
			},
			"CreatePurchaseOrder": {
				"properties": {
					"purchaseRequestId": {
						"type": "number",
						"format": "double"
					},
					"otherDetails": {
						"type": "string"
					},
					"fulfillmentDate": {
						"type": "string",
						"format": "date-time"
					},
					"vendorName": {
						"type": "string"
					}
				},
				"required": [
					"purchaseRequestId",
					"otherDetails",
					"fulfillmentDate",
					"vendorName"
				],
				"type": "object"
			},
			"CreateVoucherOrOrder": {
				"properties": {
					"orderData": {
						"$ref": "#/components/schemas/CreatePurchaseOrder"
					},
					"expenses": {
						"items": {
							"$ref": "#/components/schemas/ExpenseAttr"
						},
						"type": "array"
					},
					"chargeId": {
						"type": "number",
						"format": "double"
					},
					"requestedDate": {
						"type": "string",
						"format": "date-time"
					},
					"requestedBy": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"expenses",
					"chargeId",
					"requestedDate",
					"requestedBy",
					"description"
				],
				"type": "object"
			},
			"ApproveVoucherOrOrder": {
				"properties": {
					"disbursements": {
						"items": {
							"$ref": "#/components/schemas/DisbursementAttr"
						},
						"type": "array"
					},
					"codes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"purchaseOrderId": {
						"type": "number",
						"format": "double"
					},
					"purchaseRequestId": {
						"type": "number",
						"format": "double"
					},
					"voucherId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"codes"
				],
				"type": "object"
			},
			"RejectPurchaseOrder": {
				"properties": {
					"rejectedBy": {
						"type": "number",
						"format": "double"
					},
					"comments": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"rejectedBy",
					"comments",
					"id"
				],
				"type": "object"
			},
			"CancelPurchaseOrder": {
				"properties": {
					"cancelledBy": {
						"type": "number",
						"format": "double"
					},
					"comments": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"cancelledBy",
					"comments",
					"id"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"bearer": {
				"type": "apiKey",
				"in": "header",
				"name": "authorization"
			}
		}
	},
	"info": {
		"title": "luminous-api",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/property/getAll": {
			"get": {
				"operationId": "GetAllProperties",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PropertyAttr"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/property/{id}": {
			"get": {
				"operationId": "GetProperty",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PropertyAttr"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/property/getPropertyAssignments/{propertyId}": {
			"get": {
				"operationId": "GetPropertyAssignments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PropertyAssignmentAttr"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "propertyId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/property/getAssignedProperties/{profileId}": {
			"get": {
				"operationId": "GetAssignedProperties",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PropertyAssignmentAttr"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "profileId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/property/create": {
			"post": {
				"operationId": "CreateProperty",
				"responses": {
					"201": {
						"description": "created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PropertyAttr"
								}
							}
						}
					},
					"400": {
						"description": "bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EntityError"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PropertyAttr"
							}
						}
					}
				}
			}
		},
		"/api/property/updatePropertyStatus/{id}": {
			"patch": {
				"operationId": "UpdatePropertyStatus",
				"responses": {
					"204": {
						"description": "no content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordStatus"
						}
					}
				]
			}
		},
		"/api/property/updateProperty/{id}": {
			"patch": {
				"operationId": "UpdateProperty",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PropertyAttr"
								}
							}
						}
					},
					"400": {
						"description": "bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EntityError"
								}
							}
						}
					},
					"404": {
						"description": "not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PropertyAttr"
							}
						}
					}
				}
			}
		},
		"/api/property/updatePropertyAssignments/{id}": {
			"patch": {
				"operationId": "UpdatePropertyAssignments",
				"responses": {
					"204": {
						"description": "no content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"type": "number",
									"format": "double"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/api/property/getPaymentHistory/{propertyId}/{year}": {
			"get": {
				"operationId": "GetPaymentHistory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PaymentHistoryView"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "propertyId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "year",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/property/getTransactionHistory/{propertyId}/{year}": {
			"get": {
				"operationId": "GetTransactionHistory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PropertyTransactionHistoryView"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "propertyId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "year",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/purchase-order/getPurchaseOrder/{id}": {
			"get": {
				"operationId": "GetPurchaseOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PurchaseOrderAttr"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/purchase-order/getAllPurchaseOrdersByChargeAndStatus/{chargeId}/{status}": {
			"get": {
				"operationId": "GetAllPurchaseOrdersByChargeAndStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PurchaseOrderAttr"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "chargeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "status",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RequestStatus"
						}
					}
				]
			}
		},
		"/api/purchase-order/postPurchaseOrder": {
			"post": {
				"operationId": "PostPurchaseOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateVoucherOrOrder"
							}
						}
					}
				}
			}
		},
		"/api/purchase-order/updatePurchaseOrder/{id}": {
			"patch": {
				"operationId": "UpdatePurchaseOrder",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateVoucherOrOrder"
							}
						}
					}
				}
			}
		},
		"/api/purchase-order/approvePurchaseOrder": {
			"post": {
				"operationId": "ApprovePurchaseOrder",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ApproveVoucherOrOrder"
							}
						}
					}
				}
			}
		},
		"/api/purchase-order/rejectPurchaseOrder": {
			"post": {
				"operationId": "RejectPurchaseOrder",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RejectPurchaseOrder"
							}
						}
					}
				}
			}
		},
		"/api/purchase-order/cancelPurchaseOrder": {
			"post": {
				"operationId": "CancelPurchaseOrder",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CancelPurchaseOrder"
							}
						}
					}
				}
			}
		},
		"/api/purchase-order/notifyPurchaseOrderApprovers/{id}": {
			"post": {
				"operationId": "NotifyPurchaseOrderApprovers",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/purchase-order/postPurchaseOrderDisbursement/{id}": {
			"post": {
				"operationId": "PostPurchaseOrderDisbursement",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DisbursementAttr"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}